refactor(gui): split session and chat tile orchestration
