+++ gui/src/components/workspace-board/ChatTile.tsx
@@ -6,12 +6,6 @@ import {
   useState,
 } from "react";
 import {
-  PlusIcon,
-  RotateCcwIcon,
-  SquareTerminalIcon,
-  XIcon,
-} from "lucide-react";
-import {
   DockviewReact,
   positionToDirection,
   type DockviewApi,
@@ -19,35 +13,21 @@ import {
   type IDockviewHeaderActionsProps,
   type DockviewReadyEvent,
   type IDockviewPanel,
-  type IDockviewPanelHeaderProps,
-  type IDockviewPanelProps,
 } from "dockview-react";
 
-import { ConversationPanel } from "@/components/ConversationPanel";
-import { ConversationDockviewHeaderActions } from "@/components/conversation-panel/ConversationDockviewHeaderActions";
 import { ConversationDockviewTab } from "@/components/conversation-panel/ConversationDockviewTab";
-import { Button } from "@/components/ui/Button";
-import { ButtonGroup } from "@/components/ui/ButtonGroup";
-import {
-  Dialog,
-  DialogContent,
-  DialogDescription,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle,
-} from "@/components/ui/Dialog";
 import * as desktopClient from "@/services/desktop/client";
 
-import { AuxiliaryPaneShell } from "./AuxiliaryPaneShell";
-import { ChangesPane } from "./ChangesPane";
-import { PlaceholderPane } from "./PlaceholderPane";
-import { TerminalPane } from "./TerminalPane";
+import {
+  AuxiliaryPaneTab,
+  ChatTileAuxiliaryPane,
+  ChatTileConversationPane,
+  ChatTileHeaderActions,
+  CloseTerminalsDialog,
+} from "./ChatTileChrome";
 import {
   CHANGES_PANE_ID,
-  CHAT_PANE_ID,
-  createTerminalSessionId,
   dispatchPaneCloseRequest,
-  dispatchTerminalRestartRequest,
   INNER_CHAT_COMPONENT,
   INNER_PLACEHOLDER_COMPONENT,
   isTerminalPanelId,
@@ -67,10 +47,6 @@ import {
 } from "./chatTileLayout";
 import { useDockviewLayoutPersistence } from "./hooks/useDockviewLayoutPersistence";
 import { useDockviewTheme } from "./hooks/useDockviewTheme";
-import type {
-  PlaceholderPaneParams,
-  TerminalPaneParams,
-} from "./types/layout";
 import type { ChatTileProps } from "./types/workspaceBoard";
 import {
   animatePaneClose,
@@ -80,42 +56,6 @@ import {
 } from "./paneResizeAnimation";
 import { applySavedConversationLayout } from "./utils/savedLayout";
 import "./chat-tile-dockview.css";
-
-function AuxiliaryPaneTab({
-  api,
-  params,
-}: IDockviewPanelHeaderProps<PlaceholderPaneParams>) {
-  const isTerminal = params.kind === "terminal";
-
-  return (
-    <div className="terminal-pane-tab group/tab inline-flex h-6 min-w-0 items-center gap-1.5 px-1.5 text-xs font-medium tracking-tight text-foreground">
-      {isTerminal ? (
-        <SquareTerminalIcon className="size-3.5 shrink-0 text-muted-foreground" />
-      ) : null}
-      <span className="truncate">{params.label}</span>
-      {isTerminal ? (
-        <button
-          type="button"
-          aria-label={`Close ${params.label}`}
-          className="ml-0.5 inline-flex size-4 shrink-0 items-center justify-center rounded text-muted-foreground opacity-0 transition-all hover:bg-muted hover:text-foreground focus-visible:opacity-100 group-hover/tab:opacity-100"
-          onPointerDown={(event) => {
-            // Stop dockview from treating the close click as a tab drag/activate.
-            event.stopPropagation();
-          }}
-          onClick={(event) => {
-            event.stopPropagation();
-            dispatchPaneCloseRequest(`${params.conversationId}:${api.id}`);
-          }}
-        >
-          <XIcon className="size-3" />
-        </button>
-      ) : null}
-    </div>
-  );
-}
-
-const headerActionsClassName =
-  "relative z-20 ml-auto flex h-6 shrink-0 items-center gap-1.5 pointer-events-auto";
 
 export function ChatTile({
   binding,
@@ -384,157 +324,49 @@ export function ChatTile({
     scheduleInnerLayoutSave();
   }, [binding, scheduleInnerLayoutSave, withProgrammaticMutation]);
 
-  const headerActionsComponent = useCallback(function InnerHeaderActions({
-    activePanel,
-  }: IDockviewHeaderActionsProps) {
-    if (activePanel?.id === CHAT_PANE_ID) {
+  const headerActionsComponent = useCallback(
+    function InnerHeaderActions(props: IDockviewHeaderActionsProps) {
       return (
-        <ConversationDockviewHeaderActions
+        <ChatTileHeaderActions
+          {...props}
           binding={binding}
           canCloseChat={canCloseChat}
-          onCloseChat={onCloseChat}
-          onOpenPreview={() => {
-            handleOpenPane("preview");
-          }}
-          onOpenTerminal={handleToggleTerminal}
-          onOpenChanges={handleToggleChanges}
           isChangesOpen={openPanes.changes}
           isTerminalOpen={openPanes.terminal}
+          onAddTerminal={handleAddTerminal}
+          onCloseChat={onCloseChat}
+          onOpenChanges={handleToggleChanges}
+          onOpenPane={handleOpenPane}
+          onToggleTerminal={handleToggleTerminal}
         />
       );
-    }
-
-    const panelParams = activePanel?.params as Partial<PlaceholderPaneParams> | undefined;
-    const kind = panelParams?.kind;
-
-    if (
-      activePanel == null ||
-      (kind !== "preview" && kind !== "terminal" && kind !== "changes")
-    ) {
-      return null;
-    }
-
-    const handleClose = () => {
-      dispatchPaneCloseRequest(`${binding.conversationId}:${activePanel.id}`);
-    };
-
-    if (kind === "preview" || kind === "changes") {
-      return (
-        <div className={headerActionsClassName}>
-          <Button
-            variant="outline"
-            size="icon-sm"
-            aria-label={kind === "changes" ? "Close changes" : "Close preview"}
-            onClick={handleClose}
-          >
-            <XIcon />
-          </Button>
-        </div>
-      );
-    }
-
-    const resolvedParams = panelParams ?? {};
-    const workspacePath = typeof resolvedParams.workspacePath === "string"
-      ? resolvedParams.workspacePath
-      : "";
-    const conversationId = typeof resolvedParams.conversationId === "string"
-      ? resolvedParams.conversationId
-      : "";
-    const terminalId = createTerminalSessionId(
-      {
-        workspacePath,
-        conversationId,
-      },
-      resolvedParams.terminalKey,
-    );
-
-    return (
-      <div className={headerActionsClassName}>
-        <Button
-          variant="outline"
-          size="icon-sm"
-          aria-label="New terminal"
-          onClick={handleAddTerminal}
-        >
-          <PlusIcon />
-        </Button>
-        <ButtonGroup aria-label="Terminal actions">
-          <Button
-            variant="outline"
-            size="icon-sm"
-            aria-label="Restart terminal"
-            onClick={() => {
-              dispatchTerminalRestartRequest(terminalId);
-            }}
-          >
-            <RotateCcwIcon />
-          </Button>
-          <Button
-            variant="outline"
-            size="icon-sm"
-            aria-label="Close terminal"
-            onClick={handleClose}
-          >
-            <XIcon />
-          </Button>
-        </ButtonGroup>
-      </div>
-    );
-  }, [
-    binding,
-    canCloseChat,
-    handleAddTerminal,
-    handleOpenPane,
-    handleToggleChanges,
-    handleToggleTerminal,
-    onCloseChat,
-    openPanes.changes,
-    openPanes.terminal,
-  ]);
+    },
+    [
+      binding,
+      canCloseChat,
+      handleAddTerminal,
+      handleOpenPane,
+      handleToggleChanges,
+      handleToggleTerminal,
+      onCloseChat,
+      openPanes.changes,
+      openPanes.terminal,
+    ],
+  );
 
   const components = useMemo(
     () => ({
       [INNER_CHAT_COMPONENT]: function ConversationPane() {
         return (
-          <ConversationPanel
+          <ChatTileConversationPane
             binding={binding}
             canCloseChat={canCloseChat}
             onCloseChat={onCloseChat}
-            onOpenPreview={() => {
-              handleOpenPane("preview");
-            }}
-            onOpenTerminal={() => {
-              handleOpenPane("terminal");
-            }}
-            showHeader={false}
-            windowDragEnabled={false}
+            onOpenPane={handleOpenPane}
           />
         );
       },
-      [INNER_PLACEHOLDER_COMPONENT]: function AuxiliaryPane(
-        props: IDockviewPanelProps<PlaceholderPaneParams>,
-      ) {
-        const kind = props.params.kind;
-        const content =
-          kind === "terminal" ? (
-            <TerminalPane
-              {...(props as IDockviewPanelProps<TerminalPaneParams>)}
-            />
-          ) : kind === "changes" ? (
-            <ChangesPane {...props} />
-          ) : (
-            <PlaceholderPane {...props} />
-          );
-
-        return (
-          <AuxiliaryPaneShell
-            api={props.api}
-            paneId={`${props.params.conversationId}:${props.api.id}`}
-          >
-            {content}
-          </AuxiliaryPaneShell>
-        );
-      },
+      [INNER_PLACEHOLDER_COMPONENT]: ChatTileAuxiliaryPane,
     }),
     [binding, canCloseChat, handleOpenPane, onCloseChat],
   );
@@ -743,40 +575,11 @@ export function ChatTile({
         tabComponents={tabComponents}
         theme={theme}
       />
-      <Dialog
-        open={closeTerminalsCount > 0}
-        onOpenChange={(open) => {
-          if (!open) {
-            setCloseTerminalsCount(0);
-          }
-        }}
-      >
-        <DialogContent className="max-w-md">
-          <DialogHeader>
-            <DialogTitle>Close {closeTerminalsCount} terminals?</DialogTitle>
-            <DialogDescription>
-              This closes every terminal tab and ends its shell session. This
-              can't be undone.
-            </DialogDescription>
-          </DialogHeader>
-          <DialogFooter>
-            <Button
-              type="button"
-              variant="outline"
-              onClick={() => setCloseTerminalsCount(0)}
-            >
-              Cancel
-            </Button>
-            <Button
-              type="button"
-              variant="destructive"
-              onClick={handleCloseAllTerminalsConfirm}
-            >
-              Close terminals
-            </Button>
-          </DialogFooter>
-        </DialogContent>
-      </Dialog>
+      <CloseTerminalsDialog
+        count={closeTerminalsCount}
+        onCancel={() => setCloseTerminalsCount(0)}
+        onConfirm={handleCloseAllTerminalsConfirm}
+      />
     </section>
   );
 }
